fix(retention): route tombstones through agent write door (PAN-3357)
